refactor: 카테고리 타입으로 조회하는 로직 수정
- repository에 카테고리 타입으로 검색하는 메소드인 findByType 추가
- enums 패키지 추가
