+++ src/main/java/com/poortorich/category/repository/CategoryRepository.java
@@ -1,13 +1,16 @@
 package com.poortorich.category.repository;
 
 import com.poortorich.category.entity.Category;
+import com.poortorich.category.entity.enums.CategoryType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
 public interface CategoryRepository extends JpaRepository<Category, Long> {
 
     Optional<Category> findByName(String name);
+    List<Category> findByType(CategoryType type);
 }
